+++ otj-server-reactive/pom.xml
@@ -59,7 +59,6 @@
         <dependency>
             <groupId>com.opentable</groupId>
             <artifactId>otj-httpheaders</artifactId>
-            <version>0.1.2</version>
         </dependency>
 
         <dependency>
@@ -105,11 +104,6 @@
         <dependency>
             <groupId>io.projectreactor</groupId>
             <artifactId>reactor-core</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
         </dependency>
 
         <dependency>
